final submission with some changes
It has all README all comments working version for all tests.
There are some changes because we find out that there is a design that in some cells it has 2 output pins so this version does not work.
I made to parse more than one functions but it does not separate connections on two output gatepins.
Needs to fix it !!!!